separated relevant hpp files #6
